requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pyyaml>=6.0
